- removed tvOS scheme - using single index.js instead of platform dependent
